\name{seraphim-package}
\alias{seraphim-package}
\alias{seraphim}
\docType{package}
\title{
A package to analyse continuous phylogeographic reconstructions and conduct hypotheses testing based on landscape phylogeographic analyses.
}
\description{
The package "seraphim" (for "studying environmental rasters and phylogenetic informed movements"; Dellicour \emph{et al}. 2016a) can, for instance, be used to investigate the impact of environmental factors on the \emph{diffusion velocity} of lineages for which the dispersal history has been inferred by a continuous phylogeographic analysis. In this analytical procedure initially described in Dellicour (\emph{et al}. 2016b), functions of the R package "seraphim" can be used to (i) extract the spatio-temporal information embedded in spatiotemporally-referenced trees inferred, for instance, by the continuous phylogeographic method implemented in BEAST (Lemey \emph{et al}. 2010), (ii) compute the environmental distances associated with each phylogenetic branch for given environmental rasters, (iii) estimate the correlation between the dispersal duration and these environmental distances associated with each branch, and (iv) estimate statistical supports for these correlations using null models generated by randomisation procedures (Dellicour \emph{et al}. 2017). In addition, the package "seraphim" can be used to display the outcome of a continuous phylogeographic reconstruction, estimate a series of lineage dispersal statistics (Dellicour \emph{et al}. 2024), investigate the impact of environmental factors on the \emph{dispersal frequency} (Dellicour \emph{et al}. 2018) and \emph{dispersal location/direction} (Dellicour \emph{et al}. 2019) of lineages. Finally, the package also includes different different functions to conduct phylogeographic simulations either along phylogenetic trees (Dellicour \emph{et al}. 2018) or with a birth-death process (Dellicour \emph{et al}. 2024 and \emph{in prep.}).
}
\author{
Simon Dellicour, Rebecca Rose, Nuno Faria, Philippe Lemey, Oliver G. Pybus
}
\references{
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016a). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
Dellicour S, Rose R, Pybus OG (2016b). Explaining the geographic spread of emerging epidemics: a framework for comparing viral phylogenies and environmental landscape data. \emph{BMC Bioinformatics} 17: 82. \cr \cr
Dellicour S, Rose R, Faria NR, Vieira LFP, Bourhy H, Gilbert M, Lemey P, Pybus OG (2017). Using viral gene sequences to compare and explain the heterogeneous spatial dynamics of virus epidemics. \emph{Molecular Biology \& Evolution}, \emph{in press}. \cr \cr
Dellicour S, Baele G, Dudas G, Faria NR, Pybus OG, Suchard M, Rambaut A, Lemey P (2018). Phylodynamic assessment of intervention strategies for the West African Ebola virus outbreak. \emph{Nature Communications} 9: 2222. \cr \cr
Dellicour S, Bastide P, Rocu P, Fargette D, Hardy OJ, Suchard MA, Guindon S, Lemey P (2024). How fast are viruses spreading in the wild? \emph{PLoS Biology} 22: e3002914. \cr \cr
Dellicour S, \emph{et al.} (\emph{in prep.}). Peaks and valleys of landscape phylogeography. \cr \cr
Lemey P, Rambaut A, Welch JJ, Suchard MA (2010). Phylogeography takes a relaxed random walk in continuous space and time. \emph{Molecular Biology & Evolution} 27: 1877-1885.
}
\seealso{
\url{https://beast.community/} \cr
\url{http://www.circuitscape.org} \cr
\url{https://github.com/Circuitscape/Circuitscape.jl}
}
