\name{treesRandomisation}
\alias{treesRandomisation}
\title{
A function to randomise the position of pmhylogenetic tree branches
}
\description{
This function randomises the position of phylogenetic branches on the study area and is based on the "extraction" tables previously generated by the "treeExtractions" or "postTreeExtractions" function, which both summarise each phylogenetic branch as a movement vector. Four distinct randomisations have been implemented (see below) and, in the case of two of them, can be conducted while considering or not the influence of an environmental raster on the dispersal locations of lineages.
}
\usage{
treesRandomisation(localTreesDirectory = "",
                   nberOfExtractionFiles = 1,
                   envVariable,
                   randomProcedure = 3,
                   repulsion = NULL,
                   overwrite = FALSE,
                   showingPlots = FALSE)                   
}
\arguments{
  \item{localTreesDirectory}{ directory where the tree extraction matrices are saved (see the "treeExtractions" or "postTreeExtractions" function). }
  \item{nberOfExtractionFiles}{ number of annotated phylogenetic trees from which spatio-temporal information has been extracted. }
  \item{envVariable}{ environmental raster to consider and on which the tree branches randomisation will be conducted. If the "repulsion" parameter is set to "NULL", this raster is only used to prevent that randomised node positions fall in non-accessible areas, i.e. raster cells associated with a "NA" value (e.g., sea areas). If the "repulsion" parameter is set to "TRUE" or "FALSE", the environmental raster will influence the tree branch randomisation process, higher environmental values tending to repulse ("TRUE") or attract ("FALSE") the dispersal locations of lineages (see below for further detail). }
  \item{randomProcedure}{ randomisation procedure to use: "3" (randomisations of nodes position while maintaining the branch lengths, the tree topology, and the location of the most ancestral node), "4" (randomisations of node positions while maintaining the branches length and the location of the starting node of each branch), "5" (randomisations of node positions while maintaining the branch lengths and the location of the ending node of each branch), or "6" (randomisations of nodes position while maintaining only the branch lengths). }
  \item{repulsion}{ boolean variable specifying if the environmental raster should be treated as an environmental variable repulsing ("TRUE") or attracting ("FALSE") the dispersal location of the branches (see below for further detail), or if its environmental values should not be considered in the tree branches randomisation process ("NULL"). In the latter case, the environmental raster is only used to prevent that randomised node positions fall in non-accessible areas, i.e. raster cells associated with a "NA" value (e.g., sea areas). }
  \item{overwrite}{ boolean variable specifying if existing simulation outputs should be overwritten or not (if set to "TRUE" and if output files already exists for the simulation n°\emph{i}, this simulation will not be conducted and the function will directly conduct simulation \emph{i}+1). }
  \item{showingPlots}{ boolean variable specifying if visualisations the different randomisations have to be displayed (and here automatically saved) or not. }
}
\value{
In practice, the function generates a table containing the spatio-tremporal information corresponding to each randomised tree, and which are similar to the tables generated by the "treeExtractions" or "postTreeExtractions" function.  
}
\details{
Several randomisation procedures of tree branches have been implemented: (i) a randomisation of the branch positions while maintaining their length, the tree topology, and the location of the most ancestral node (procedure n°3), (ii) a randomisation of the branch positions while maintaining their length and the location of the starting node of each branch (procedure n°4), (iii) a randomisation of the branch positions while maintaining their length and the location of the ending node of each branch (procedure n°5), and (iv) a randomisation of the branches position while maintaining only their length  (procedure n°6). The latter procedure can be useful, e.g., when analysing non phylogenetically informed movement data like capture-mark-recapture movement data. Note that in each case, branch positions are independently randomised for each tree. Note that randomisation procedures n°1 (using already performed tree randomisations) and n°2 (sing already performed RRW simulations) are not implemented in this function and correspond to procedures solely implemented in the "spreadFactors" function. \cr \cr
If the "repulsion" parameter is set to "TRUE" or "FALSE", the environmental raster will influence the tree branch randomisation process, higher environmental values tending to repulse ("TRUE") or attract ("FALSE") the dispersal locations of lineages. In practice, each branch position will be randomised 100 times and the environmental values at all the potential randomised child (or "ending") node locations are recorded. Those environmental values \eqn{V} are then used to specify the probability \eqn{p_j} to randomly pick a potential randomised location \eqn{j} given the environmental value \eqn{v_j} at that location: \eqn{p_j=(max(V)-v_j)/(max(V)-min(V))} if "repulsion" is set to "TRUE" (the environmental raster tending to attract the dispersal location of lineages) or \eqn{p_j=(v_j-min(v))/(max(V)-min(V))} if "repulsion" is set to "FALSE" (the environmental raster tending to repulse the dispersal location of lineages). Important note: because only relevant in those cases, these repulsion/attraction randomisation settings are not available for the randomisation procedures n°3 and 4 (if the randomisation procedure n°5 or 6 is selected, the "repulsion" parameter will be set to "NULL" by the fonction). }
}
\author{
Simon Dellicour
}
\references{
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016a). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
Dellicour S, Rose R, Pybus OG (2016b). Explaining the geographic spread of emerging epidemics: a framework for comparing viral phylogenies and environmental landscape data. \emph{BMC Bioinformatics} 17: 82.
}
