\name{simulatorBRW}
\alias{simulatorBRW}
\title{
A function to simulate a Bownian (BRW) or relaxed (RRW) random walk diffusion process
}
\description{
This function simulates a Bownian (BRW) or relaxed (RRW) random walk diffusion process while preventing dispersal in inaccessible raster cells.
}
\usage{
simulatorRRW2(RRW = TRUE,
			  envVariable =
raster(matrix(c(runif(600,5,10),runif(1000,0,5)),nrow=40,ncol=40)),
			  sigma = 1,
			  ancestPosition = c(0.4,0.5),
			  birthRate = 0.2,
			  samplingRate = 0.2,
			  startingYear = 0,
			  samplingWindow = c(10,50),
			  timeSlice = 0.1,
			  timeIntervale = 1,
			  showingPlots = FALSE)
}
\arguments{
  \item{RRW}{ boolean variable specifying if a Bownian (BRW) or relaxed (RRW) random walk diffusion process should be simulated. }
  \item{envVariable}{ raster that will determine the environmental heterogeneity (cells with a NaN value will be considered as inaccessible). }

  \item{sigma}{ standard deviation of the normal distribution with mean parameter set to zero in which longitudinal and latitudinal displacement values are randomly picked (if "RRW" is set to "FALSE", i.e. in the case of a BRW simulation), or the value of the scale parameter of the Cauchy distribution with location parameter set to zero in which longitudinal and latitudinal displacement values are randomly picked (if "RRW" is set to "TRUE", i.e. in the case of a RRW simulation (see also the note below).)
  \item{ancestPosition}{ vector of geographic coordinates (longitude, latitude) of the most ancestral node position, i.e. the starting position of the RRW diffusion simulation. }
  \item{birthRate}{ value defining the rate at which lineages are splited in two new lineages (expressed in events per lineage per time unit). }
  \item{samplingRate}{ value defining the rate at which lineage are sampled and thus stopped moving and splitting on the raster (expressed in events per lineage per time unit). }
  \item{startingYear}{ year (or moment in another time unit) at which the simulation begins from the ancestral node position. }
  \item{samplingWindow}{ vector of time values (in the same time unit as the "startingYear") defining the starting and ending times of the sampling period, i.e. the period during which the sampling rate will be effective. }
  \item{timeSlice}{ time interval (in the same time unit as the "startingYear") at which unsampled lineages perform a new movement on the raster. }
  \item{timeIntervale}{ time interval (in the same time unit as the "startingYear") at which the plot displaying the simulation is updated (only taken into account when showingPlots=TRUE). }
  \item{showingPlots}{ boolean variable specifying if the different plots have to be displayed or not. }
}
\value{
The function simulates a BRW or RRW diffusion process while preventing dispersal in inaccessible raster cells. In practice, the function generates a list of two objects: (i) a table containing the spatio-tremporal information of the simulated BRW/RRW diffusion, with one line per lineage/phylogenetic branch. Such a table similar to those generated by the "treeExtractions" function. (ii) A tree object of class "phylo" that thus summarizes the phylogenetic relationships among simulated lineages.
}
\note{
At each time step, once the longitudinal and latitudinal displacements have been independently drawn in a normal (BRW simulation) or Cauchy distribution (RRW simulation), the resulting movement vector is then randomly rotated around the position of the evolving lineage at the previous time step to avoid that sporadic long-distance dispersal events solely occure along a longitudinal or latitudinal gradient (a rotation step that is thus mostly relevant for RRW simulations in which such long-distance dispersal events are more frequent due to the nature of the Cauchy distribution).}
\references{
Dellicour S, Rose R, Faria N, Lemey P, Pybus OG (2016). SERAPHIM: studying environmental rasters and phylogenetically-informed movements. \emph{Bioinformatics} 32 (20): 3204-3206. \cr \cr
}
\author{
Simon Dellicour
}
